Changed appriate data types

+++ kernel_module/src/tdisk_plugin.h
@@ -35,7 +35,7 @@ inline static loff_t plugin_get_size(const char *plugin)
   * This function writes the given bytes to
   * plugin at the given position. It also measures the performance
  **/
-inline static int plugin_write_data(const char *plugin, void *data, loff_t pos, unsigned int length, struct device_performance *perf)
+inline static int plugin_write_data(const char *plugin, void *data, loff_t pos, size_t length, struct device_performance *perf)
 {
 	int ret;
 
@@ -60,7 +60,7 @@ inline static int plugin_write_data(const char *plugin, void *data, loff_t pos, 
   * This function reads the given bytes from
   * plugin at the given position. It also measures the performance
  **/
-inline static int plugin_read_data(const char *plugin, void *data, loff_t pos, unsigned int length, struct device_performance *perf)
+inline static int plugin_read_data(const char *plugin, void *data, loff_t pos, size_t length, struct device_performance *perf)
 {
 	int ret;
 
@@ -102,7 +102,7 @@ inline static int plugin_write_bio_vec(const char *plugin, struct bio_vec *bvec,
 	if(ret == 0)
 	{
 		(*pos) += bvec->bv_len;
-		return bvec->bv_len;
+		return (int)bvec->bv_len;
 	}
 	return ret;
 }
@@ -118,7 +118,7 @@ inline static int plugin_read_bio_vec(const char *plugin, struct bio_vec *bvec, 
 	if(ret == 0)
 	{
 		(*pos) += bvec->bv_len;
-		return bvec->bv_len;
+		return (int)bvec->bv_len;
 	}
 	return ret;
 }
